refactor: move service to gen-services
