+++ crates/gen-services/src/services/error.rs
