--- crates/persons/src/application.rs
@@ -1,1 +1,0 @@
-pub mod service;
