--- crates/persons/src/port.rs
@@ -1,1 +1,0 @@
-pub mod repository;
